# client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "English Reader 客户端接口",
    "version": "0.1.0",
    "description": "只含 /v1/client。由 scripts/export_contract.py 导出，不要手改。\n\n**只增不减**（架构铁律 5）：可以加字段，不能删字段或改字段含义。侧载安装的客户端更新滞后，手机上大概率长期跑着旧版本。\n\n**capabilities 区分「没数据」和「没实现」**：一个 null 字段本身是歧义的，`root_known: null` 可能是「这个词没有词根」，也可能是「水平估计还没做」，而客户端要把这两种画成不同的样子。\n\n**幂等键重复上报算正常，不算错误**：离线客户端会把拿不准的都重发一遍。\n\n**今日包不含真题**：真题走 /library?source=cet4|cet6|kaoyan。"
  },
  "paths": {
...
    "/v1/client/events": {
      "post": {
        "tags": [
          "client:reading"
        ],
        "summary": "批量上报交互事件",
        "description": "Accepts duplicates by design — a retry is the protocol working.",
        "operationId": "report_events_v1_client_events_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventBatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventBatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "client:reading"
        ],
        "summary": "按序号往后取事件（多设备同步用）",
        "description": "这个学习者的事件，序号大于 ``after`` 的那些。\n\n**P9 §6:同步要变双向。** 在这之前只有上报——一台设备把事件送上来，\n而另一台设备永远看不到它。多设备要一个会合点，这就是那个会合点的读取口。\n\n**序号是 `client_events.id`，不是新造的东西。** 那张表从 P2 起就是\n``AUTOINCREMENT``，它一直是「服务端收到即分配的单调序号」；\n再造一个会得到第二个顺序，然后两个顺序说反话。\n**服务端只存不解释**（架构铁律 2 的后半句），所以 `payload` 原样回去。\n\n**拉回自己推上去的事件是正常的。** 游标是「大于某个号」，而自己的事件也在\n那个号后面。客户端按 `idem_key` 认出来并跳过——**这比让服务端按设备过滤好**:\n按设备过滤要服务端知道「哪台设备产生了哪条」，而设备换了令牌就不认了，\n那时它会以为自己的历史不存在。\n\n**GET 而不是 POST，`after` 在查询串里**:它是一次读取，缓存与重试的语义\n都该按读取来。",
        "operationId": "event_feed_v1_client_events_get",
        "parameters": [
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "After"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 500,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventFeedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/client/reviews/spelling": {
      "post": {
        "tags": [
          "client:review"
        ],
        "summary": "上报一次拼写",
        "description": "Recorded, never scheduled on.\n\n决定 13 keeps spelling out of the scheduler entirely: it is an optional\nreinforcement, and letting it move due dates would mix a productive-recall\ndifficulty into a recognition track. What is stored is what was typed, not\njust whether it was right — that is the whole value of the record.",
        "operationId": "report_spelling_v1_client_reviews_spelling_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpellingIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpellingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/client/reviews/spellings": {
      "post": {
        "tags": [
          "client:review"
        ],
        "summary": "批量上报拼写（离线补报用）",
        "description": "The offline path for spelling, which P4 gave answers and not this.\n\n決定 12 of P4 added a batch endpoint for answers because \"地铁里每答一题都\n失败\" broke 架构前提 2 for half the product. Spelling was out of scope that\nday and kept its one-at-a-time online POST, so the last step of the day was\nstill the one step that needed a network. Same store, same idempotency key,\nsame per-item verdict as the answers batch — a client that already speaks\nthat one needs no new vocabulary here.\n\n**Order does not matter here**, unlike answers: each attempt is an\nindependent record, and spelling never touches the scheduler (决定 13). They\nare applied in the order given anyway, because replaying in sequence is\nwhat the client's outbox does.\n\nRe-applying an attempt whose first try failed writes a second row in\n`spelling_attempts`. That is a duplicate record rather than a wrong one —\nnothing reads the table to make a decision — and the alternative is the\nattempt being silently dropped.",
        "operationId": "report_spellings_v1_client_reviews_spellings_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpellingsIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpellingsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/client/today": {
      "get": {
        "tags": [
          "client:today"
        ],
        "summary": "今日包：当天的文章与复习，一次拿全",
        "description": "One response the client can work from all day without a network.\n\nNew endpoint rather than fields added to an existing one — 架构铁律 5 lets\nus add, never change, and P2 deliberately left this shape unguessed\n(`phase-2.html` §6) rather than bolting an invented one onto `/articles`.\n\n**Does not include exam papers** (决定 9). They have their own entrance at\n``/v1/client/library?source=cet4|cet6|kaoyan``. The response says so in\n``excludes_exam_papers`` so a client cannot conclude otherwise by accident.\n\n**条件请求（2026-09-16 加）。** 这个包实测约 1 MB，而客户端每次开复习那一格\n都要它——绝大多数时候内容跟手机上那份**一模一样**，却照样过一遍隧道\n（实测 0.85–1.4 秒，流量也是真金白银）。带 ``If-None-Match`` 来、内容没变，\n就回 304 和零字节。\n\n**加的是一个响应头和一条分支，不是新字段**，所以老客户端一行不用改：\n它不发 ``If-None-Match``，就永远走 200 那条路（铁律 5）。\n\n哈希算的是**序列化之后的响应体**，不是它的某几个字段——任何一处变化都算变化，\n包括复习进度、文章读到哪。宁可多发一次，也不能把变了的说成没变。",
        "operationId": "today_v1_client_today_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodayResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClientEvent": {
        "properties": {
          "idem_key": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Idem Key",
            "description": "客户端生成的幂等键"
          },
          "type": {
            "type": "string",
            "title": "Type",
            "description": "事件类型"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "occurred_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurred At",
            "description": "客户端时钟，可能不准"
          }
        },
        "type": "object",
        "required": [
          "idem_key",
          "type"
        ],
        "title": "ClientEvent",
        "description": "One reported interaction.\n\n``idem_key`` is generated by the client, not the server: the client is the\nonly party that knows whether this is the same event it failed to upload ten\nminutes ago on a train."
      },
...
      "EventBatch": {
        "properties": {
          "events": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ClientEvent"
                },
                {
                  "$ref": "#/components/schemas/UnusableEvent"
                }
              ]
            },
            "type": "array",
            "maxItems": 500,
            "title": "Events"
          }
        },
        "type": "object",
        "title": "EventBatch"
      },
      "EventBatchResponse": {
        "properties": {
          "accepted": {
            "type": "integer",
            "title": "Accepted"
          },
          "duplicates": {
            "type": "integer",
            "title": "Duplicates"
          },
          "failed": {
            "type": "integer",
            "title": "Failed"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/EventResult"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "accepted",
          "duplicates",
          "failed",
          "results"
        ],
        "title": "EventBatchResponse",
        "description": "Duplicates are the protocol working, not a fault.\n\nAn offline client retries whatever it is unsure about. What matters to the\nclient is the per-item verdict: 决定 11 says it may delete an outbox file\nonly for a key this list reports as landed — never on the strength of the\nHTTP status alone."
      },
      "EventFeedResponse": {
        "properties": {
          "learner": {
            "$ref": "#/components/schemas/Learner"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/LoggedEventOut"
            },
            "type": "array",
            "title": "Events"
          },
          "through": {
            "type": "integer",
            "title": "Through",
            "description": "这一批里最大的序号。**一条都没有时回传你给的那个 after**，这样客户端不用区分「空」和「到底了」"
          },
          "latest": {
            "type": "integer",
            "title": "Latest",
            "description": "服务端手上最大的序号。`through < latest` 就是还有"
          },
          "more": {
            "type": "boolean",
            "title": "More",
            "description": "还有没有下一段"
          }
        },
        "type": "object",
        "required": [
          "learner",
          "events",
          "through",
          "latest",
          "more"
        ],
        "title": "EventFeedResponse",
        "description": "按序号往后取一段事件。"
      },
      "EventResult": {
        "properties": {
          "idem_key": {
            "type": "string",
            "title": "Idem Key"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "accepted 收下了 / duplicate 重复 / failed 处理失败 / rejected 不认识"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "idem_key",
          "status"
        ],
        "title": "EventResult"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SpellingIn": {
        "properties": {
          "item_key": {
            "type": "string",
            "title": "Item Key"
          },
          "typed": {
            "type": "string",
            "title": "Typed"
          }
        },
        "type": "object",
        "required": [
          "item_key",
          "typed"
        ],
        "title": "SpellingIn"
      },
...
      "SpellingResponse": {
        "properties": {
          "correct": {
            "type": "boolean",
            "title": "Correct"
          },
          "expected": {
            "type": "string",
            "title": "Expected"
          }
        },
        "type": "object",
        "required": [
          "correct",
          "expected"
        ],
        "title": "SpellingResponse",
        "description": "Recorded, never scheduled on.\n\nSpelling stays out of the scheduler entirely: it is optional reinforcement,\nand letting it move due dates would mix a productive-recall difficulty into\na recognition track. What is stored is what was *typed*, not just whether it\nwas right — that is the whole value of the record."
      },
...
      "SpellingsIn": {
        "properties": {
          "spellings": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/SpellingItem"
                },
                {
                  "$ref": "#/components/schemas/UnusableItem"
                }
              ]
            },
            "type": "array",
            "maxItems": 500,
            "title": "Spellings"
          }
        },
        "type": "object",
        "title": "SpellingsIn",
        "description": "**元素同样是联合类型**，理由见 :class:`UnusableItem`。\n\n这一路还没出过事，而形状和作答那一路一模一样:一条字段验不过去就整批 422，\n而这个端点的契约写的也是「the same per-item verdict as the answers batch」。\n**等它出事再改，就是等一次「手机一直转圈」**——那一次已经付过了。"
      },
      "SpellingsResponse": {
        "properties": {
          "accepted": {
            "type": "integer",
            "title": "Accepted"
          },
          "duplicates": {
            "type": "integer",
            "title": "Duplicates"
          },
          "failed": {
            "type": "integer",
            "title": "Failed"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/SpellingResult"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "accepted",
          "duplicates",
          "failed",
          "results"
        ],
        "title": "SpellingsResponse",
        "description": "离线补报的一批拼写。\n\nNo progress block, unlike the answers batch: spelling happens after the\nday's review is already finished, so there is nothing left for it to move."
      },
...
      "TodayResponse": {
        "properties": {
          "learner": {
            "$ref": "#/components/schemas/Learner"
          },
          "capabilities": {
            "$ref": "#/components/schemas/Capabilities"
          },
          "day": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Day",
            "description": "复习那一天的日期"
          },
          "articles": {
            "items": {
              "$ref": "#/components/schemas/ArticleResponse"
            },
            "type": "array",
            "title": "Articles",
            "description": "今天的正课。**正好 3 篇**，由配置项 today_article_count 决定"
          },
          "extra_articles": {
            "items": {
              "$ref": "#/components/schemas/ArticleResponse"
            },
            "type": "array",
            "title": "Extra Articles",
            "description": "**永远是空的。**加餐已于 2026-09-12 取消——往期本身就是加餐，往列表下面翻就有。字段留着不删是因为铁律 5 只增不减；把 today_extra_count 改回非零就能恢复"
          },
          "reviews": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReviewDayResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "**P9 起不再下发。** 复习那一份由设备自己算（句子来自 `/v1/client/sentences`，状态来自它重放自己的事件日志）。字段留着不删——铁律 5 只增不减，而老客户端还在读它；但服务端组装它意味着**每次请求都要建会话、入队**，也就是服务端在写学习状态，而那正是那条线禁止的事"
          },
          "settings": {
            "$ref": "#/components/schemas/TodaySettings"
          },
          "excludes_exam_papers": {
            "type": "boolean",
            "title": "Excludes Exam Papers",
            "description": "**今日包 ≠ 今天能读的全部。**真题走自己的入口 /library?source=cet4|cet6|kaoyan。写在响应里而不只是写在文档里——一个以为这就是全部的客户端会静默地藏起 452 篇真题"
          }
        },
        "type": "object",
        "required": [
          "learner",
          "capabilities",
          "articles",
          "extra_articles",
          "settings",
          "excludes_exam_papers"
        ],
        "title": "TodayResponse",
        "description": "当天的文章与复习，一次拿全。\n\n**全文内联，不是清单**（P4 决定 8）。实测一篇约 300 KB，三篇不到 1 MB——\n清单式的话联不上网就只剩清单，那就不叫今日包了。"
      },
...
      "UnusableEvent": {
        "properties": {
          "idem_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idem Key",
            "description": "能认出是哪一条就够了，其余字段不做要求"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "UnusableEvent",
        "description": "一条验不过去的上报。**留着它，好让它有资格被逐条拒绝。**\n\n2026-09-19 加，起因是复习作答那一路的同一个形状（见\n`review/routes.py` 的 `UnusableItem`）:这个端点回的是**逐条**裁决\n（`EventBatchResponse` 的注释写着「What matters to the client is the\nper-item verdict」），而 pydantic 的校验发生在 handler 之前——\n一条字段验不过去就让整批 500 条一起 422，逐条那套完全没机会跑。\n\n这一路实际发作过一次，形状还不一样:客户端一批发 534 条而这里写着\n`max_length=500`，于是**每次都是 422，队列只会变长**（手机卡了一天）。\n那一次是客户端改成分片修的；**而「一条坏的不许拖垮整批」是这一侧的事**。"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}